Updated files to compile
Updated files to conform to the latest Coding Guidelines
