+++ addons/binding/org.openhab.binding.enphaseenvoy/build.properties
@@ -7,4 +7,5 @@ bin.includes=META-INF/,\
              lib/okhttp-digest-1.17.jar,\
              OSGI-INF/,\
              ESH-INF/,\
-             NOTICE
+             NOTICE,\
+			 about.html
