+++ addons/binding/org.openhab.binding.enphaseenvoy/pom.xml
@@ -7,12 +7,12 @@
 	<parent>
 		<artifactId>pom</artifactId>
 		<groupId>org.openhab.binding</groupId>
-		<version>2.3.0-SNAPSHOT</version>
+		<version>2.5.0-SNAPSHOT</version>
 	</parent>
 
 	<groupId>net.hentschel.openhab</groupId>
 	<artifactId>org.openhab.binding.enphaseenvoy</artifactId>
 	<packaging>eclipse-plugin</packaging>
 	<name>EnphaseEnvoy Binding</name>
-	<version>0.0.1</version>
+	
 </project>
