+++ addons/binding/org.openhab.binding.enphaseenvoy/src/main/java/org/openhab/binding/enphaseenvoy/discovery/internal/EnphaseEnvoyDiscoveryParticipant.java
@@ -1,7 +1,12 @@
 /**
+ * Copyright (c) 2010-2018 by the respective copyright holders.
  *
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License v1.0
+ * which accompanies this distribution, and is available at
+ * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.openhab.binding.enphaseenvoy.discovery;
+package org.openhab.binding.enphaseenvoy.discovery.internal;
 
 import java.net.Inet4Address;
 import java.util.HashMap;
