+++ addons/binding/org.openhab.binding.enphaseenvoy/src/main/java/org/openhab/binding/enphaseenvoy/discovery/internal/EnphaseEnvoyDiscoveryService.java
@@ -1,4 +1,12 @@
-package org.openhab.binding.enphaseenvoy.discovery;
+/**
+ * Copyright (c) 2010-2018 by the respective copyright holders.
+ *
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License v1.0
+ * which accompanies this distribution, and is available at
+ * http://www.eclipse.org/legal/epl-v10.html
+ */
+package org.openhab.binding.enphaseenvoy.discovery.internal;
 
 import java.io.IOException;
 import java.util.Collections;
@@ -17,10 +25,14 @@ import org.eclipse.smarthome.core.thing.ThingUID;
 import org.openhab.binding.enphaseenvoy.internal.EnphaseEnvoyBindingConstants;
 import org.openhab.binding.enphaseenvoy.internal.EnphaseEnvoyBridgeConfiguration;
 import org.openhab.binding.enphaseenvoy.internal.EnphaseEnvoyBridgeHandler;
-import org.openhab.binding.enphaseenvoy.protocol.InverterProduction;
+import org.openhab.binding.enphaseenvoy.protocol.internal.InverterProduction;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+/**
+ * @author thomashentschel
+ *
+ */
 public class EnphaseEnvoyDiscoveryService extends AbstractDiscoveryService implements ExtendedDiscoveryService {
 
     private static final Logger logger = LoggerFactory.getLogger(EnphaseEnvoyDiscoveryService.class);
