+++ addons/binding/org.openhab.binding.enphaseenvoy/src/main/java/org/openhab/binding/enphaseenvoy/internal/EnphaseEnvoyBridgeHandler.java
@@ -1,14 +1,10 @@
 /**
- * Copyright (c) 2014,2018 by the respective copyright holders.
+ * Copyright (c) 2010-2018 by the respective copyright holders.
  *
- * See the NOTICE file(s) distributed with this work for additional
- * information regarding copyright ownership.
- *
- * This program and the accompanying materials are made available under the
- * terms of the Eclipse Public License 2.0 which is available at
- * http://www.eclipse.org/legal/epl-2.0
- *
- * SPDX-License-Identifier: EPL-2.0
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License v1.0
+ * which accompanies this distribution, and is available at
+ * http://www.eclipse.org/legal/epl-v10.html
  */
 package org.openhab.binding.enphaseenvoy.internal;
 
@@ -38,9 +34,9 @@ import org.eclipse.smarthome.core.thing.ThingTypeUID;
 import org.eclipse.smarthome.core.thing.ThingUID;
 import org.eclipse.smarthome.core.thing.binding.BaseBridgeHandler;
 import org.eclipse.smarthome.core.types.Command;
-import org.openhab.binding.enphaseenvoy.discovery.EnphaseEnvoyDiscoveryService;
-import org.openhab.binding.enphaseenvoy.protocol.InverterProduction;
-import org.openhab.binding.enphaseenvoy.protocol.SystemProduction;
+import org.openhab.binding.enphaseenvoy.discovery.internal.EnphaseEnvoyDiscoveryService;
+import org.openhab.binding.enphaseenvoy.protocol.internal.InverterProduction;
+import org.openhab.binding.enphaseenvoy.protocol.internal.SystemProduction;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
