--- addons/binding/org.openhab.binding.enphaseenvoy/src/main/java/org/openhab/binding/enphaseenvoy/protocol/SystemProduction.java
@@ -1,16 +1,0 @@
-/**
- *
- */
-package org.openhab.binding.enphaseenvoy.protocol;
-
-/**
- * @author thomashentschel
- *
- */
-public class SystemProduction {
-
-    public long wattHoursToday;
-    public long wattHoursSevenDays;
-    public long wattHoursLifetime;
-    public long wattsNow;
-}
