+++ addons/binding/org.openhab.binding.enphaseenvoy/src/main/java/org/openhab/binding/enphaseenvoy/protocol/internal/SystemProduction.java
@@ -1,0 +1,21 @@
+/**
+ * Copyright (c) 2010-2018 by the respective copyright holders.
+ *
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License v1.0
+ * which accompanies this distribution, and is available at
+ * http://www.eclipse.org/legal/epl-v10.html
+ */
+package org.openhab.binding.enphaseenvoy.protocol.internal;
+
+/**
+ * @author thomashentschel
+ *
+ */
+public class SystemProduction {
+
+    public long wattHoursToday;
+    public long wattHoursSevenDays;
+    public long wattHoursLifetime;
+    public long wattsNow;
+}
